Fix #51, add monitor table
Replace the "free space" table with a more general disk monitoring
table, which can have entries for volume free space (which is what was
there) as well as an estimate of directory usage.

Table and TLM was renamed accordingly.  This changes the definition of
the TLM report to include the extra info, so it requires a ground system
update.